#166 Fix ClassCastException on Android (#196)
Due to Dart not always transporting the 
callback handle as a `long` the forced cast to 
`Long` on Android fails if the handle is only an 
`int`.
Props to @mbabker for finding this out.

As a solution, the number is cast to `Number` 
instead. The actual value can still be extracted
as a `long`.